+++ src/TuioNet/Tuio11/Tuio11Cursor.cs
@@ -39,7 +39,7 @@ namespace TuioNet.Tuio11
             {
                 var message = new OSCMessage("/tuio/2Dcur");
                 message.Append("set");
-                message.Append(SessionId);
+                message.Append((int)SessionId);
                 message.Append(Position.X);
                 message.Append(Position.Y);
                 message.Append(Velocity.X);
